# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /targets:
    get:
      operationId: listTargets
      tags: [targets]
      summary: Enerji hedeflerini listele (ilerleme dahil)
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnergyTargetWithProgress"
    post:
      operationId: createTarget
      tags: [targets]
      summary: Enerji hedefi ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnergyTargetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnergyTarget"
  /targets/{id}:
    patch:
      operationId: updateTarget
      tags: [targets]
      summary: Enerji hedefini güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnergyTargetUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnergyTarget"
    delete:
      operationId: deleteTarget
      tags: [targets]
      summary: Enerji hedefini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── RISKS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    EnergyTarget:
      type: object
      required: [id, name, baselineYear, targetYear, targetReductionPercent, createdAt]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EnergyTargetWithProgress:
      type: object
      required: [id, name, baselineYear, targetYear, targetReductionPercent, createdAt, yearlyProgress]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        baselineKwh:
          type: ["number", "null"]
        yearlyProgress:
          type: array
          items:
            type: object
            required: [year]
            properties:
              year:
                type: integer
              actualKwh:
                type: ["number", "null"]
              reductionPercent:
                type: ["number", "null"]
    EnergyTargetInput:
      type: object
      required: [name, baselineYear, targetYear, targetReductionPercent]
      properties:
        unitId:
          type: integer
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: string
    EnergyTargetUpdate:
      type: object
      properties:
        unitId:
          type: integer
        name:
          type: string
        baselineYear:
          type: integer
        targetYear:
          type: integer
        targetReductionPercent:
          type: number
        notes:
          type: string
...